--- anchor/common/message_validator/src/validation.rs
@@ -1,116 +1,0 @@
-use ssv_types::message::SignedSSVMessage;
-use crate::validation::Result::Accept;
-
-// TODO taken from go-SSV as rough guidance. feel free to adjust as needed. https://github.com/ssvlabs/ssv/blob/e12abf7dfbbd068b99612fa2ebbe7e3372e57280/message/validation/errors.go#L55
-pub enum ValidationFailure {
-    WrongDomain,
-    NoShareMetadata,
-    UnknownValidator,
-    ValidatorLiquidated,
-    ValidatorNotAttesting,
-    EarlySlotMessage,
-    LateSlotMessage,
-    SlotAlreadyAdvanced,
-    RoundAlreadyAdvanced,
-    DecidedWithSameSigners,
-    PubSubDataTooBig(usize),
-    IncorrectTopic,
-    NonExistentCommitteeID,
-    RoundTooHigh,
-    ValidatorIndexMismatch,
-    TooManyDutiesPerEpoch,
-    NoDuty,
-    EstimatedRoundNotInAllowedSpread,
-    EmptyData,
-    MismatchedIdentifier,
-    SignatureVerification,
-    PubSubMessageHasNoData,
-    MalformedPubSubMessage,
-    NilSignedSSVMessage,
-    NilSSVMessage,
-    SSVDataTooBig,
-    InvalidRole,
-    UnexpectedConsensusMessage,
-    NoSigners,
-    WrongRSASignatureSize,
-    ZeroSigner,
-    SignerNotInCommittee,
-    DuplicatedSigner,
-    SignerNotLeader,
-    SignersNotSorted,
-    InconsistentSigners,
-    InvalidHash,
-    FullDataHash,
-    UndecodableMessageData,
-    EventMessage,
-    UnknownSSVMessageType,
-    UnknownQBFTMessageType,
-    InvalidPartialSignatureType,
-    PartialSignatureTypeRoleMismatch,
-    NonDecidedWithMultipleSigners,
-    DecidedNotEnoughSigners,
-    DifferentProposalData,
-    MalformedPrepareJustifications,
-    UnexpectedPrepareJustifications,
-    MalformedRoundChangeJustifications,
-    UnexpectedRoundChangeJustifications,
-    NoPartialSignatureMessages,
-    NoValidators,
-    NoSignatures,
-    SignersAndSignaturesWithDifferentLength,
-    PartialSigOneSigner,
-    PrepareOrCommitWithFullData,
-    FullDataNotInConsensusMessage,
-    TripleValidatorIndexInPartialSignatures,
-    ZeroRound,
-    DuplicatedMessage,
-    InvalidPartialSignatureTypeCount,
-    TooManyPartialSignatureMessages,
-    EncodeOperators,
-}
-
-impl From<&ValidationFailure> for Result {
-    fn from(value: &ValidationFailure) -> Self {
-        match value {
-            ValidationFailure::WrongDomain
-            | ValidationFailure::NoShareMetadata
-            | ValidationFailure::UnknownValidator
-            | ValidationFailure::ValidatorLiquidated
-            | ValidationFailure::ValidatorNotAttesting
-            | ValidationFailure::EarlySlotMessage
-            | ValidationFailure::LateSlotMessage
-            | ValidationFailure::SlotAlreadyAdvanced
-            | ValidationFailure::RoundAlreadyAdvanced
-            | ValidationFailure::DecidedWithSameSigners
-            | ValidationFailure::PubSubDataTooBig(_)
-            | ValidationFailure::IncorrectTopic
-            | ValidationFailure::NonExistentCommitteeID
-            | ValidationFailure::RoundTooHigh
-            | ValidationFailure::ValidatorIndexMismatch
-            | ValidationFailure::TooManyDutiesPerEpoch
-            | ValidationFailure::NoDuty
-            | ValidationFailure::EstimatedRoundNotInAllowedSpread => Result::Ignore,
-            _ => Result::Reject,
-        }
-    }
-}
-
-pub struct Validator;
-
-pub enum Error {}
-
-pub enum Result {
-    Accept,
-    Reject,
-    Ignore,
-}
-
-impl Validator {
-    pub(crate) fn validate(
-        &mut self,
-        _message: SignedSSVMessage,
-    ) -> std::result::Result<Result, Error> {
-
-        Ok(Accept)
-    }
-}
